+++ resources/views/neura/context-menu/separator.blade.php
@@ -1,3 +1,3 @@
-<div {{ $attributes->merge(['class' => 'h-px bg-separator my-1']) }}></div>
+<div {{ $attributes->merge(['class' => 'h-px bg-black/[0.05] dark:bg-white/[0.06] my-1 -mx-1']) }}></div>