+++ resources/views/neura/dropdown/separator.blade.php
@@ -1,1 +1,1 @@
-<div {{ $attributes->merge(['class' => 'my-1 -mx-(--dropdown-padding) h-px grid-cols-subgrid col-span-full bg-separator']) }}></div>
+<div {{ $attributes->merge(['class' => 'my-1 -mx-(--dropdown-padding) h-px grid-cols-subgrid col-span-full bg-black/[0.05] dark:bg-white/[0.06]']) }}></div>